+++ examples/autoreload.py
@@ -1,10 +1,9 @@
-from imp import reload
 import os,sys; sys.path.insert(0,os.path.dirname(os.path.dirname(__file__)))
 
 from htag import Tag
 
 """
-AUTORELOAD is possible with this two runners : WebHTTP, BrowserStarletteHTTP & BrowserStarletteWS
+AUTORELOAD is possible with this three runners : WebHTTP, BrowserStarletteHTTP & BrowserStarletteWS
 (could be very handy in development phase)
 
 See this example, on how to instanciate the runner and use uvicorn/autoreload
@@ -15,7 +14,7 @@ class Demo(Tag.div):
 
     def init(self):
         for c in range(2000000,16581375,10000):
-            self <= Tag.button("Hello",_style=f"background:#{hex(c)[2:]}")
+            self <= Tag.button("Hi",_style=f"background:#{hex(c)[2:]}")
 
 #############################################################################################
 
@@ -27,7 +26,7 @@ app = BrowserStarletteHTTP( Demo )
 
 if __name__=="__main__":
     import uvicorn
-    uvicorn.run("autoreload:app",host="127.0.0.1",port=8000,reload=True,factory=True)
+    uvicorn.run("autoreload:app",host="127.0.0.1",port=8000,reload=True)
 
-    # or the classic :
-    # app.run()
+    ## or the classic :
+    #app.run()