+++ examples/camshot.py
@@ -2,6 +2,9 @@ import os,sys; sys.path.insert(0,os.path.dirname(os.path.dirname(__file__)))
 from htag import Tag
 
 class Cam(Tag.video):
+    """ Htag component to start the cam, and to take screnshot (by clicking on it)"""
+
+    # some javascripts needed for the component
     statics = Tag.script("""
 function startCam(video) {
     navigator.mediaDevices.getUserMedia({
@@ -13,36 +16,44 @@ function startCam(video) {
         video.play();
     })
     .catch(function(err) {
-        alert("An error occurred: " + err);
+        const o = document.createElement("div");
+        o.innerHTML = "NO CAMERA: " + err;
+        video.replaceWith(o);
     });
 }
 
-function takeCamShot(video,width,height) {
+function takeCamShot(video) {
+    let {width, height} = video.srcObject.getTracks()[0].getSettings();
+
     let canvas = document.createElement("canvas");
     let context = canvas.getContext('2d');
     canvas.width = width;
     canvas.height = height;
-    context.drawImage(video, 0, 0, width, height);
+    context.drawImage(video, 0, 0, canvas.width, canvas.height);
     return canvas.toDataURL('image/jpeg');
 }
 """)
 
+    # js which will be executed, each time the component is appended
     js = """startCam(tag);"""
 
-    def init(self,callback=None,width=320,height=240):
+    def init(self,callback=None,width=300,height=300):
+        self.width=width
+        self.height=height
         self["style"]=f"width:{width}px;height:{height}px;border:1px solid black"
         if callback:
-            self["onclick"]=self.bind( callback, bytes("takeCamShot(this,%s,%s)" % (width,height),"utf8") )
+            self["onclick"]=self.bind( callback, b"takeCamShot(this)" )
 
 class App(Tag.body):
+    """ An Htag App to handle the camera, screenshots are displayed in the flow"""
+
     statics = b"function error(m) {alert(m)}"
 
     def init(self):
-        self <= Tag.p("Click on the video, to take a screenshot ;-)")
         self <= Cam(self.takeShot)
 
     def takeShot(self,o,dataurl):
-        self <= Tag.img(_src=dataurl)
+        self <= Tag.img(_src=dataurl,_style="max-width:%spx;max-height:%spx;" % (o.width,o.height))