+++ htag/runners/browserstarlettehttp.py
@@ -17,9 +17,11 @@ from starlette.responses import HTMLResponse,JSONResponse
 from starlette.routing import Route
 
 
-class BrowserStarletteHTTP:
+class BrowserStarletteHTTP(Starlette):
     """ Simple ASync Web Server (with starlette) with HTTP interactions with htag.
         Open the rendering in a browser tab.
+
+        The instance is an ASGI htag app
     """
     def __init__(self,tagClass:type):
         assert issubclass(tagClass,Tag)
@@ -34,6 +36,11 @@ window.addEventListener('DOMContentLoaded', start );
 
         self.renderer=HRenderer(tagClass, js, lambda: os._exit(0))
 
+        Starlette.__init__(self,debug=True, routes=[
+            Route('/', self.GET, methods=["GET"]),
+            Route('/', self.POST, methods=["POST"]),
+        ])
+
     async def GET(self,request) -> HTMLResponse:
         return HTMLResponse( str(self.renderer) )
 
@@ -42,20 +49,10 @@ window.addEventListener('DOMContentLoaded', start );
         dico = await self.renderer.interact(data["id"],data["method"],data["args"],data["kargs"])
         return JSONResponse(dico)
 
-
-    def __call__(self):
-        """ create a uvicorn factory/asgi, to make it compatible with uvicorn
-            from scratch.
-        """
-        return Starlette(debug=True, routes=[
-            Route('/', self.GET, methods=["GET"]),
-            Route('/', self.POST, methods=["POST"]),
-        ])
-
     def run(self, host="127.0.0.1", port=8000, openBrowser=True):   # localhost, by default !!
         import uvicorn,webbrowser
         if openBrowser:
             webbrowser.open_new_tab(f"http://{host}:{port}")
 
-        uvicorn.run(self(), host=host, port=port)
+        uvicorn.run(self, host=host, port=port)